Patches the flaky tests for GetRulesAmendments:
The cause is that the enacting rules tests test create rules in the same game that the getrulesamendments tests later uses to verify the endpoint can find a game without any rules or amendments. If the getrulesamendments tests run before enacting, all tests pass. Otherwise, the one test in getrulesamendments fails.